refactor: Accept a lot of IDE suggestions

+++ src/main/java/org/sqlite/core/CorePreparedStatement.java
@@ -126,11 +126,11 @@ public abstract class CorePreparedStatement extends JDBC4Statement {
 
             case REAL:
                 // long to Julian date
-                batch(pos, Double.valueOf((value / 86400000.0) + 2440587.5));
+                batch(pos, value / 86400000.0 + 2440587.5);
                 break;
 
             default: // INTEGER:
-                batch(pos, Long.valueOf(value / config.getDateMultiplier()));
+                batch(pos, value / config.getDateMultiplier());
         }
     }
 }
